requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pymysql>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
flask-cors>=4.0.0
flask-bcrypt>=1.0.1
flask-jwt-extended>=4.5.3
flask-mail>=0.9.1
redislite>=6.0
